+++ client/src/components/Containers/TrustBadgeContainer.js
@@ -1,14 +1,6 @@
-import React, { useState, useEffect } from 'react';
-import {
-  Grid,
-  Typography,
-  Card,
-  CardMedia,
-  useMediaQuery,
-  Slide,
-} from '@material-ui/core';
-import { AiFillGithub, AiFillLinkedin } from 'react-icons/ai';
-import { useTheme } from '@material-ui/core/styles';
+import React from 'react';
+import { Grid, Typography } from '@material-ui/core';
+import Carousel from 'components/Carousel/Carousel';
 import useStyles from './TrustBadgeContainerStyles';
 
 const defaultTestimonials = [
@@ -40,54 +32,7 @@ const defaultTestimonials = [
 
 const TrustBadgeContainer = () => {
   const classes = useStyles(useStyles);
-  const theme = useTheme();
-  const mobile = useMediaQuery(theme.breakpoints.down('xs'));
-  const tablet = useMediaQuery(theme.breakpoints.down('sm'));
-
-  // const [testimonials, setTestimonials] = useState(defaultTestimonials);
   const testimonials = defaultTestimonials;
-  const [active, setActive] = useState({
-    next: 0,
-    prev: testimonials.length - 1,
-  });
-  const { next, prev } = active;
-
-  const transitionSlide = (idx) => {
-    if (next === idx) {
-      return { position: 'absolute', transitionTimingFunction: 'linear' };
-    }
-    if (prev === idx) {
-      return {
-        position: 'absolute',
-        transitionTimingFunction: 'linear',
-      };
-    }
-    return { visibility: 'hidden' };
-  };
-
-  const changeActive = () => {
-    if (!mobile) return;
-    const timer = setTimeout(() => {
-      const nextValue = next + 1 >= defaultTestimonials.length ? 0 : next + 1;
-      setActive({ next: nextValue, prev: next });
-    }, 300);
-    // eslint-disable-next-line consistent-return
-    return () => {
-      clearTimeout(timer);
-    };
-  };
-
-  useEffect(() => {
-    if (!mobile) return;
-    const timer = setTimeout(() => {
-      const nextValue = next + 1 >= defaultTestimonials.length ? 0 : next + 1;
-      setActive({ next: nextValue, prev: next });
-    }, 4000);
-    // eslint-disable-next-line consistent-return
-    return () => {
-      clearTimeout(timer);
-    };
-  }, [mobile, next]);
 
   return (
     <Grid
@@ -119,51 +64,8 @@ const TrustBadgeContainer = () => {
         xs={12}
         spacing={3}
       >
-        {/* Map Over Grid Item & Testimonials */}
-        {testimonials.map((testimonial, idx) => (
-          <Slide
-            direction={next === idx ? 'left' : 'right'}
-            in={mobile ? next === idx : true}
-            style={mobile ? transitionSlide(idx) : ''}
-            timeout={mobile ? { enter: 800, exit: 600 } : 0}
-          >
-            <Grid
-              onClick={changeActive}
-              key={testimonial.name.replace(' ', '')}
-              item
-              md={3}
-              sm={6}
-              xs={12}
-            >
-              <Card classes={{ root: classes.card }}>
-                <CardMedia
-                  className={classes.media}
-                  image={testimonial.src}
-                  title={testimonial.name}
-                />
-                <Typography variant="h6" align="center">
-                  {testimonial.name}
-                </Typography>
-                <span>
-                  <a
-                    href={testimonial.github}
-                    target="_blank"
-                    rel="noopener noreferrer"
-                  >
-                    <AiFillGithub color="black" size={tablet ? 20 : 35} />
-                  </a>
-                  <a
-                    href={testimonial.linkedin}
-                    target="_blank"
-                    rel="noopener noreferrer"
-                  >
-                    <AiFillLinkedin color="black" size={tablet ? 20 : 35} />
-                  </a>
-                </span>
-              </Card>
-            </Grid>
-          </Slide>
-        ))}
+        {/* Custom Carousel Component for Testimonials Object */}
+        <Carousel testimonials={testimonials} />
       </Grid>
     </Grid>
   );
